feat(develop): Fixed the field rename's bug

+++ src/akita.rs
@@ -126,7 +126,7 @@ impl AkitaMapper for Akita {
         let columns = T::fields();
         let enumerated_columns = columns
             .iter().filter(|f| f.exist)
-            .map(|c| format!("`{}`", c.name))
+            .map(|c| format!("`{}`", &c.alias.to_owned().unwrap_or(c.name.to_string())))
             .collect::<Vec<_>>()
             .join(", ");
         let select_fields = wrapper.get_select_sql();
@@ -160,7 +160,7 @@ impl AkitaMapper for Akita {
         let columns = T::fields();
         let enumerated_columns = columns
             .iter().filter(|f| f.exist)
-            .map(|c| format!("`{}`", c.name))
+            .map(|c| format!("`{}`", &c.alias.to_owned().unwrap_or(c.name.to_string())))
             .collect::<Vec<_>>()
             .join(", ");
         let select_fields = wrapper.get_select_sql();
@@ -191,7 +191,7 @@ impl AkitaMapper for Akita {
         let col_len = columns.len();
         let enumerated_columns = columns
             .iter().filter(|f| f.exist)
-            .map(|c| format!("`{}`", c.name))
+            .map(|c| format!("`{}`", c.alias.to_owned().unwrap_or(c.name.to_string())))
             .collect::<Vec<_>>()
             .join(", ");
         let mut conn = self.acquire()?;
@@ -201,10 +201,10 @@ impl AkitaMapper for Akita {
         }) {
             let sql = match conn {
                 #[cfg(feature = "akita-mysql")]
-                DatabasePlatform::Mysql(_) => format!("SELECT {} FROM {} WHERE `{}` = ? limit 1", &enumerated_columns, &table.complete_name(), &field.name),
+                DatabasePlatform::Mysql(_) => format!("SELECT {} FROM {} WHERE `{}` = ? limit 1", &enumerated_columns, &table.complete_name(), &field.alias.to_owned().unwrap_or(field.name.to_string())),
                 #[cfg(feature = "akita-sqlite")]
-                DatabasePlatform::Sqlite(_) => format!("SELECT {} FROM {} WHERE `{}` = ${} limit 1", &enumerated_columns, &table.complete_name(), &field.name, col_len + 1),
-                _ => format!("SELECT {} FROM {} WHERE `{}` = ${} limit 1", &enumerated_columns, &table.complete_name(), &field.name, col_len + 1),
+                DatabasePlatform::Sqlite(_) => format!("SELECT {} FROM {} WHERE `{}` = ${} limit 1", &enumerated_columns, &table.complete_name(), &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
+                _ => format!("SELECT {} FROM {} WHERE `{}` = ${} limit 1", &enumerated_columns, &table.complete_name(), &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
             };
 
             let rows = conn.execute_result(&sql, (id.to_value(),).into())?;
@@ -227,7 +227,7 @@ impl AkitaMapper for Akita {
         let columns = T::fields();
         let enumerated_columns = columns
             .iter().filter(|f| f.exist)
-            .map(|c| format!("`{}`", c.name))
+            .map(|c| format!("`{}`", c.alias.to_owned().unwrap_or(c.name.to_string())))
             .collect::<Vec<_>>()
             .join(", ");
         let select_fields = wrapper.get_select_sql();
@@ -310,10 +310,10 @@ impl AkitaMapper for Akita {
         }) {
             let sql = match conn {
                 #[cfg(feature = "akita-mysql")]
-                DatabasePlatform::Mysql(_) => format!("delete from {} where `{}` = ?", &table.name, &field.name),
+                DatabasePlatform::Mysql(_) => format!("delete from {} where `{}` = ?", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string())),
                 #[cfg(feature = "akita-sqlite")]
-                DatabasePlatform::Sqlite(_) => format!("delete from {} where `{}` = ${}", &table.name, &field.name, col_len + 1),
-                _ => format!("delete from {} where `{}` = ${}", &table.name, &field.name, col_len + 1),
+                DatabasePlatform::Sqlite(_) => format!("delete from {} where `{}` = ${}", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
+                _ => format!("delete from {} where `{}` = ${}", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
             };
             let _rows = conn.execute_result(&sql, (id.to_value(),).into())?;
             Ok(conn.affected_rows())
@@ -341,10 +341,10 @@ impl AkitaMapper for Akita {
         }) {
             let sql = match conn {
                 #[cfg(feature = "akita-mysql")]
-                DatabasePlatform::Mysql(_) => format!("delete from {} where `{}` in (?)", &table.name, &field.name),
+                DatabasePlatform::Mysql(_) => format!("delete from {} where `{}` in (?)", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string())),
                 #[cfg(feature = "akita-sqlite")]
-                DatabasePlatform::Sqlite(_) => format!("delete from {} where `{}` in (${})", &table.name, &field.name, col_len + 1),
-                _ => format!("delete from {} where `{}` = ${}", &table.name, &field.name, col_len + 1),
+                DatabasePlatform::Sqlite(_) => format!("delete from {} where `{}` in (${})", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
+                _ => format!("delete from {} where `{}` = ${}", &table.name, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
             };
             let ids = ids.iter().map(|v| v.to_value().to_string()).collect::<Vec<String>>().join(",");
             let _rows = conn.execute_result(&sql, (ids,).into())?;
@@ -379,7 +379,7 @@ impl AkitaMapper for Akita {
                 if !col.exist || col.field_type.ne(&FieldType::TableField) {
                     continue;
                 }
-                let col_name = &col.name;
+                let col_name = &col.name.to_string();
                 let mut value = data.get_obj_value(&col_name);
                 match &col.fill {
                     None => {}
@@ -428,28 +428,28 @@ impl AkitaMapper for Akita {
                     #[allow(unreachable_patterns)]
                     match conn {
                         #[cfg(feature = "akita-mysql")]
-                        DatabasePlatform::Mysql(_) => format!("`{}` = ?", &col.name),
+                        DatabasePlatform::Mysql(_) => format!("`{}` = ?", &col.alias.to_owned().unwrap_or(col.name.to_string())),
                         #[cfg(feature = "akita-sqlite")]
-                        DatabasePlatform::Sqlite(_) => format!("`{}` = ${}",&col.name, x + 1),
-                        _ => format!("`{}` = ${}", &col.name, x + 1),
+                        DatabasePlatform::Sqlite(_) => format!("`{}` = ${}",&col.alias.to_owned().unwrap_or(col.name.to_string()), x + 1),
+                        _ => format!("`{}` = ${}", &col.alias.to_owned().unwrap_or(col.name.to_string()), x + 1),
                     }
                 })
                 .collect::<Vec<_>>()
                 .join(", ");
             let sql = match conn {
                 #[cfg(feature = "akita-mysql")]
-                DatabasePlatform::Mysql(_) => format!("update {} set {} where `{}` = ?", &table.name, &set_fields, &field.name),
+                DatabasePlatform::Mysql(_) => format!("update {} set {} where `{}` = ?", &table.name, &set_fields, &field.alias.to_owned().unwrap_or(field.name.to_string())),
                 #[cfg(feature = "akita-sqlite")]
-                DatabasePlatform::Sqlite(_) => format!("update {} set {} where `{}` = ${}", &table.name, &set_fields, &field.name, col_len + 1),
-                _ => format!("update {} set {} where `{}` = ${}", &table.name, &set_fields, &field.name, col_len + 1),
+                DatabasePlatform::Sqlite(_) => format!("update {} set {} where `{}` = ${}", &table.name, &set_fields, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
+                _ => format!("update {} set {} where `{}` = ${}", &table.name, &set_fields, &field.alias.to_owned().unwrap_or(field.name.to_string()), col_len + 1),
             };
             let mut values: Vec<Value> = Vec::with_capacity(columns.len());
-            let id = data.get_obj_value(&field.name);
+            let id = data.get_obj_value(&field.name.to_string());
             for col in columns.iter() {
                 if !col.exist || col.field_type.ne(&FieldType::TableField) {
                     continue;
                 }
-                let col_name = &col.name;
+                let col_name = &col.name.to_string();
                 let mut value = data.get_obj_value(col_name);
                 match &col.fill {
                     None => {}
@@ -494,7 +494,7 @@ impl AkitaMapper for Akita {
         for entity in entities.iter() {
             for col in columns.iter().filter(|col| col.exist ) {
                 let data = entity.to_value();
-                let mut value = data.get_obj_value(&col.name);
+                let mut value = data.get_obj_value(&col.name.to_string());
                 match &col.fill {
                     None => {}
                     Some(v) => {
@@ -529,7 +529,8 @@ impl AkitaMapper for Akita {
         let data = entity.to_value();
         let mut values: Vec<Value> = Vec::with_capacity(columns.len());
         for col in columns.iter().filter(|col| col.exist ) {
-            let mut value = data.get_obj_value(&col.name);
+            println!("dataS:{:?}", data);
+            let mut value = data.get_obj_value(&col.name.to_string());
             match &col.fill {
                 None => {}
                 Some(v) => {
@@ -570,7 +571,7 @@ impl AkitaMapper for Akita {
             FieldType::TableId(_) => true,
             FieldType::TableField => false,
         }) {
-            data.get_obj_value(&field.name).unwrap_or(&Value::Nil)
+            data.get_obj_value(&field.name.to_string()).unwrap_or(&Value::Nil)
         } else { &Value::Nil };
         match id {
             Value::Nil => {
